+++ packages/rm-menu/CHANGELOG.md
@@ -1,6 +1,22 @@
 # Changelog
 All notable changed to this project will be documented in this file
 
+## [1.0.2] - 2020-11-30
+- replace `focus-trap` with `@riot-material/focus-manager`
+- do great rewrite
+- split js logic from style and html
+- update README
+
+## [any] - Unreleased
+- better handling when menu is not anchored
+- add documentation for "opening" event
+- add documentation for "open" event
+- add documentation for "closing" event
+- add documentation for "close" event
+- add documentation for "wantsclose" event
+- scroll at highlighted
+- highlight on scroll
+
 ## [1.0.1] - 2020-11-26
 - update dependencies
 - move source in `src` folder and bundled files in `dist`